Fixing OOM from mmap
I adapted the code in tyche-musl to implement a free.
See the commit in musl.
The gist of it is that redis apparently alloc/dealloc a bunch of pages.
In theory our current implementation provides a big enough mempool to
not crash, but as we were ignoring frees (munmap) we were stupidly
running out of memory.

I got a bitmap-based tracker for free pages from chat gpt and it seems
to mostly work. I've seen a few transient errors in the debug
environment but could not reproduce them afterwards.

TODO: we still need to make sure that this new version works inside the
tyche enclaves.